refactor: terms 관련 파일을 feature/terms 폴더 내로 이동
