+++ lib/features/terms/domain/entities/terms.dart
