+++ lib/features/terms/presentation/pages/terms_inner_page.dart
@@ -1,10 +1,9 @@
-import 'package:firebase_analytics/firebase_analytics.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_html/flutter_html.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:go_router/go_router.dart';
 import 'package:mongbi_app/core/font.dart';
-import 'package:mongbi_app/domain/entities/terms.dart';
+import 'package:mongbi_app/features/terms/domain/entities/terms.dart';
 
 class TermsDetailPage extends StatelessWidget {
   const TermsDetailPage({super.key, required this.termsList});
