+++ lib/features/terms/presentation/pages/widgets/terms_text_widget.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:mongbi_app/core/font.dart';
-import 'package:mongbi_app/core/get_responsive_ratio_by_width.dart';
 
 class TermsHeaderText extends StatelessWidget {
   const TermsHeaderText({super.key});
